Exercises - Part 2
Fixed the bug by adapting the code from SearchExercises to receive the setExercises from Home and removing the useState with the same synthax and also renaming the prop exercises to item in the {} from the const searchedExercises. Now we can access the Exercises that are being rendered as a simple <p /> from the data on the api.